Update date.js
api ui5.date.pick do not work if in datePicker control, a date is already selected with current month and same day.
for e.g if a date is already selected in datePicker which falls on 'monday' then user cannot select next monday date in current month.
it was  a logical issue in coding.